:truck: [RENAME] 폴더구조 변경 #39
